fix(runtime): address unresolved P1 review followups
